updated file names to match new class structure
